Clean up, testing
Cleaned up the loose implementation of graphing the audio: all of it is in the functions now, so it should be unneeded.

Also did some testing with a few audio files, comparing the program's output to Audacity's. There were some issues, namely with representation differences (Audacity uses dB (logarithmic), while my program appears to measure magnitude linearly. Logarithmic magnitude graphing was.. sub-standard, and figuring out a way to make it work is beyond the current combination of my skills and available time)