Rudimentary fourier analysis
Added a crude algorithm to find information about the signal: basically, it tries to find three peaks in the fourier spectrum of the data.

I'm fully aware it's very crude, but time is limited as it always is (I have to get up in three hours). I'm quite glad that it even works as well as it does, honestly.

I'll clean up everything in the morning.

ALso there's a bunch of notes at the bottom. They'll be removed before the final version, obviously.